feat(vaults): split vaults/all into job & handler
