+++ lib/db.js
@@ -37,3 +37,12 @@ module.exports.batchSet = async (table, items, batchSize = 25) => {
     }),
   );
 };
+
+module.exports.scan = async (table) => {
+  const { Items: items } = await client
+    .scan({
+      TableName: table,
+    })
+    .promise();
+  return items;
+};
